Move App components to package
