+++ src/components/App/index.ts
@@ -1,0 +1,4 @@
+import AppComponent from './app';
+export type App = typeof AppComponent;
+
+export { AppWrapper } from './app-wrapper';
