+++ src/index.tsx
@@ -10,7 +10,7 @@ import 'typeface-roboto';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import AppWrapper from './components/app-wrapper';
+import { AppWrapper } from './components/App';
 
 const container = document.getElementById('root');
 if (container) {
